Investigating of01 save crash.
The weird crash that would affect any attempt to save on the map has turned out to be a simple naming conflict: if an entity A, entity B and the movewith field of entity B share the same name, the game will fail to write a save.

What has to be done now is removing anything from the targetname fields of all glass panels that rotate with their doors.